requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiosqlite==0.19.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
websockets==12.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
manim==0.18.0
pillow==10.1.0
numpy==1.25.2
matplotlib==3.8.2
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1